Move clickToSend to its own folder
