+++ src/components/clickToSend/clickToSend.css
@@ -1,0 +1,3 @@
+.clickable {
+  cursor: pointer;
+}
