+++ src/components/send/send.css
@@ -15,7 +15,3 @@
   line-height: 14px;
   color: grey;
 }
-
-.clickable {
-  cursor: pointer;
-}
