+++ src/components/transactions/transactionType.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TooltipWrapper } from '../timestamp';
 import sytles from './transactions.css';
-import ClickToSend from '../send/clickToSend';
+import ClickToSend from '../clickToSend';
 
 const TransactionType = (props) => {
   let type;
